Change Ped Analysis to use correct field

+++ ATSPM.Application.Reports/Business/LeftTurnGapReport/PedActuationResult.cs
@@ -8,7 +8,7 @@ namespace ATSPM.Application.Reports.Business.LeftTurnGapReport
     public class PedActuationResult
     {
         public double PedActuationPercent { get; set; }
-        public int CyclesWithPedCalls { get; set; }
+        public double CyclesWithPedCalls { get; set; }
         public Dictionary<DateTime, double> PercentCyclesWithPedsList { get; set; }
         public string Direction { get; set; }
         public string OpposingDirection { get; set; }
